+++ docs/source/core/asset_manager.rst
@@ -2,7 +2,20 @@
     :class: hidden-section
 
 hawk_eye.core.asset_manager
---------------------------------
+=======================================
+
+Overview
+---------------------------------------
+This file contains functions for interacting with Google Cloud Storage
+to upload and download different assets. As of writting (2021), Image Recognition
+stores model training runs, testing dependencies, and components for data generation
+in the cloud.
+
+For data generation, new data will be added whenever there is a test flight or
+competition where data is gathered.
+
+@TODO uploading process
+
 
 .. automodule:: hawk_eye.core.asset_manager
    :members:
