+++ hawk_eye/setup/requirements-cpu.txt
@@ -1,3 +1,3 @@
 -f https://download.pytorch.org/whl/torch_stable.html
-torch==1.6.0+cpu
-torchvision==0.7.0+cpu
+torch==1.7.1+cpu
+torchvision==0.8.2+cpu
