+++ hawk_eye/setup/requirements-gpu.txt
@@ -1,3 +1,2 @@
--f https://download.pytorch.org/whl/torch_stable.html
-torch==1.7.0+cu110
-torchvision==0.8.1+cu110
+torch==1.7.1
+torchvision==0.8.2
